Move state to interface crate
