+++ program/src/lib.rs
@@ -2,7 +2,5 @@
 
 mod entrypoint;
 mod processor;
-#[allow(unused)]
-mod state;
 
 pinocchio_pubkey::declare_id!("PToken1111111111111111111111111111111111111");
