+++ program/src/processor/shared/burn.rs
@@ -1,9 +1,11 @@
 use pinocchio::{account_info::AccountInfo, program_error::ProgramError, ProgramResult};
-use token_interface::error::TokenError;
-
-use crate::{
-    processor::{check_account_owner, is_owned_by_system_program_or_incinerator, validate_owner},
+use token_interface::{
+    error::TokenError,
     state::{account::Account, mint::Mint},
+};
+
+use crate::processor::{
+    check_account_owner, is_owned_by_system_program_or_incinerator, validate_owner,
 };
 
 #[inline(always)]
