+++ program/src/processor/shared/toggle_account_state.rs
@@ -1,10 +1,10 @@
 use pinocchio::{account_info::AccountInfo, program_error::ProgramError, ProgramResult};
-use token_interface::error::TokenError;
-
-use crate::{
-    processor::validate_owner,
+use token_interface::{
+    error::TokenError,
     state::{account::Account, account_state::AccountState, mint::Mint},
 };
+
+use crate::processor::validate_owner;
 
 #[inline(always)]
 pub fn process_toggle_account_state(accounts: &[AccountInfo], freeze: bool) -> ProgramResult {
