+++ program/src/processor/ui_amount_to_amount.rs
@@ -1,9 +1,7 @@
-use core::str;
-
-use crate::state::mint::Mint;
 use pinocchio::{
     account_info::AccountInfo, program::set_return_data, program_error::ProgramError, ProgramResult,
 };
+use token_interface::state::mint::Mint;
 
 use super::{check_account_owner, try_ui_amount_into_amount};
 
@@ -12,8 +10,8 @@ pub fn process_ui_amount_to_amount(
     accounts: &[AccountInfo],
     instruction_data: &[u8],
 ) -> ProgramResult {
-    let ui_amount =
-        str::from_utf8(instruction_data).map_err(|_error| ProgramError::InvalidInstructionData)?;
+    let ui_amount = core::str::from_utf8(instruction_data)
+        .map_err(|_error| ProgramError::InvalidInstructionData)?;
 
     let mint_info = accounts.first().ok_or(ProgramError::NotEnoughAccountKeys)?;
     check_account_owner(mint_info)?;
